+++ NEWS.md
@@ -17,7 +17,7 @@
 # mvPot 0.1.3
 
 * Correction of cluster definition condition for parallel computing.
-* Addition of mvTProbQuasiMonteCarlo for efficient quasi Monte Carlo estimation of t ditribution functions.
+* Addition of mvTProbQuasiMonteCarlo for efficient quasi Monte Carlo estimation of t distribution functions.
 
 # mvPot 0.1.2