+++ NEWS.md
@@ -1,3 +1,10 @@
+# mvPot 0.1.7
+* Fix hyperlink in documentation.
+
+
+# mvPot 0.1.6
+* Fix an compilation issue on Fedora.
+
 # mvPot 0.1.5
 
 * Exposed number of replications for integration.